src/api/km-rag-function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-http-fastapi==1.0.0b1
openai==1.61.0
semantic_kernel==1.19.0
azure-search-documents==11.6.0b3
azure-ai-projects==1.0.0b5
azure-identity==1.19.0
azure-ai-inference==1.0.0b7
pyodbc==5.2.0
# httpx==0.27.2